Revert "Back commonPipelineEnvironment step by shared class (#821)"
This reverts commit eb57c8df7b9a4b7a9f44ae57f5054303a0c9f74e.
