+++ test/groovy/util/JenkinsResetDefaultCacheRule.groovy
@@ -6,7 +6,6 @@ import org.junit.runners.model.Statement
 
 import com.lesfurets.jenkins.unit.BasePipelineTest
 import com.sap.piper.DefaultValueCache
-import com.sap.piper.CommonPipelineEnvironment
 
 class JenkinsResetDefaultCacheRule implements TestRule {
 
@@ -28,7 +27,6 @@ class JenkinsResetDefaultCacheRule implements TestRule {
             @Override
             void evaluate() throws Throwable {
                 DefaultValueCache.reset()
-                CommonPipelineEnvironment.getInstance().reset()
                 base.evaluate()
             }
         }
